# specs/002-ingest-publish/contracts/jobs.openapi.yaml
openapi: 3.1.0
info:
  title: 入库任务
  version: 0.2.0
  description: |
    GeoTIFF 入库的排队、执行与查询，以及切片耗时采样。

    入库是分钟级的 GDAL 任务，因此写入 `job` 表由独立 worker 领取，不用
    FastAPI 的 `BackgroundTasks`（宪章 Technology Constraints）。
    任务的领取是原子的：条件更新 `WHERE status='queued'`，所以多个副本不会
    领到同一条。

    切片与 WMTS 的地址不在本文件，见 `tiles.openapi.yaml`。
servers:
  - url: http://localhost:8080/api
    description: 默认部署经 Caddy 网关暴露的目录 API
paths:
  /jobs:
    get:
      summary: 列出最近的入库任务
      description: 新到旧。界面上要能看见自己的上传与失败原因，所以状态与错误都原样返回。
      parameters:
        - in: query
          name: status
          description: 按状态过滤
          schema:
            type: string
            enum: [queued, running, success, failed, cancelled]
        - in: query
          name: limit
          description: 最多返回几条
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
      responses:
        "200":
          description: 任务列表
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Job" }
        "400":
          description: 未知的 status 值
  /jobs/uploads:
    post:
      summary: 将 GeoTIFF 入库排入队列
      description: |
        文件边读边写并计数，超出 `WEBGIS_MAX_UPLOAD_BYTES` 立即中止并清掉半截文件。
        扩展名不是 `.tif` / `.tiff` 时当场拒绝——这个答案在上传那一刻就已知，
        没必要先排队、等 worker 转完再报一句「不是影像文件」。
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, acquired_at]
              properties:
                file: { type: string, format: binary }
                acquired_at: { type: string, format: date-time }
      responses:
        "202":
          description: 任务已排队
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
        "400":
          description: 时间格式无效，或文件不是 .tif / .tiff
        "413":
          description: 超过大小上限
  /jobs/{job_id}:
    get:
      summary: 读取任务状态
      responses:
        "200":
          description: 任务
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
        "404":
          description: 未知任务
    delete:
      summary: 取消还在排队的任务
      description: |
        只对 `queued` 有效。已经在跑的不给取消——半途掐断只会留下状态与文件
        不一致。已经处于终态的返回 409。
      responses:
        "200":
          description: 已取消
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
        "404":
          description: 未知任务
        "409":
          description: 已处于终态，或已被 worker 领走
  /tiles/timing:
    get:
      summary: 实测一次切片耗时
      description: |
        这里**真的发起一次请求**来测量，不是估算。宪章 C8 要求：采样过才报耗时，
        观测到才报命中。

        `cache` 只有两种取值：
        - `none` —— 出口确实没有缓存层；
        - `unverified` —— 声明装了缓存层，但本端点无法观测命中率。

        永远不会返回 `hit`。命中率要从缓存层自己的日志读，不在这个端点里假设。
        不给 `url` 时 `sampled` 为 false、`duration_ms` 为 null，并说明原因——
        返回一个编造的 0 会让性能记录失真。
      parameters:
        - in: query
          name: url
          required: true
          description: |
            待采样的 COG 地址。与切片端点同一套路径校验：只放行数据目录下的
            本地 GeoTIFF，远程需在 `WEBGIS_REMOTE_COG_HOSTS` 里显式列出主机。
            （`required: true` 是就本端点的语义而言；服务端仍会校验缺失情况并返回 403。）
          schema: { type: string }
        - in: query
          name: z
          schema: { type: integer, default: 0 }
        - in: query
          name: x
          schema: { type: integer, default: 0 }
        - in: query
          name: y
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: 采样结果
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TileTiming" }
        "403":
          description: 地址不在允许范围内
components:
  schemas:
    Job:
      type: object
      required: [id, type, status, progress, created_at, updated_at]
      properties:
        id: { type: string }
        type: { type: string, enum: [ingest] }
        status:
          type: string
          enum: [queued, running, success, failed, cancelled]
          description: |
            终态是 success / failed / cancelled。进程重启时卡在 `running` 的会被
            明确标成 failed 并在 error 里写明原因，不静默丢弃。
        progress: { type: number, description: 0 到 1，目前只有起止两档 }
        error:
          type: string
          nullable: true
          description: 失败原因，用户取消时为「cancelled by user」
        payload:
          type: object
          properties:
            source_path: { type: string, description: 上传原件的落地路径 }
            acquired_at: { type: string, format: date-time, description: 拍摄时间 }
            filename: { type: string, description: 上传时的文件名 }
            bytes: { type: integer, description: 实际收到的字节数 }
            item_id:
              type: string
              nullable: true
              description: 成功后生成的目录条目 id；入库完成时取它换图层描述
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    TileTiming:
      type: object
      required: [duration_ms, cache, sampled, note]
      properties:
        duration_ms:
          type: number
          nullable: true
          description: 实测毫秒数；未采样时为 null
        cache:
          type: string
          enum: [none, unverified]
          description: 出口缓存层的状态。**不含 hit**——命中必须实测
        sampled:
          type: boolean
          description: 是否真的发起了请求
        ok:
          type: boolean
          description: 采样到的响应是否成功。sampled 只说明发过请求，4xx/5xx 时 ok 为 false。
        status:
          type: integer
          description: 被采样请求的状态码；未采样时缺省
        bytes:
          type: integer
          description: 被采样响应的字节数；未采样时缺省
        server_timing:
          type: string
          nullable: true
          description: 上游的 Server-Timing 头，可用于拆分服务端耗时
        note:
          type: string
          description: 测量条件。进程内采样与经网关采样含义不同，会写明是哪一种